pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qsense-cli"
version = "0.2.0"
description = "Minimal CLI multimodal understanding tool for agents"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "openai>=1.0",
    "httpx>=0.27",
    "jsonschema>=4.23",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
video = ["av>=12.0"]

[project.scripts]
qsense = "qsense.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
qsense = ["registry.yaml"]
